+++ nestjs/src/health/health.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 import { TerminusModule } from '@nestjs/terminus'
 import { MongooseHealthIndicator } from '@nestjs/terminus'
-import { HealthService } from '@/health/health.service'
-import { HealthController } from '@/health/health.controller'
+import { HealthService } from '@/src/health/health.service'
+import { HealthController } from '@/src/health/health.controller'
 
 @Module({
   imports: [TerminusModule],
