+++ nestjs/src/user/dto/update-user.dto.ts
@@ -1,6 +1,6 @@
 import { PartialType } from '@nestjs/mapped-types'
 import { CreateUserDto } from '@/src/user/dto/create-user.dto'
-import { IsDate, IsOptional } from '@nestjs/class-validator'
+import { IsDate, IsOptional } from 'class-validator'
 
 export class UpdateUserDto extends PartialType(CreateUserDto) {
